requirements.txt
kivy>=2.3.0
HomeAssistant-API>=4.0.0
websocket-client>=1.8.0